+++ src/App.tsx
@@ -1,10 +1,6 @@
 import { useState } from 'react';
 import * as S from './App.styles';
-import CardCVCNumber from '@/components/CardCVCNumber/CardCVCNumber';
-import CardExpirationDate from '@/components/CardExpirationDate/CardExpirationDate';
-import CardNumber from '@/components/CardNumber/CardNumber';
-import CardPreview from '@/components/CardPreview/CardPreview';
-import Spacing from '@/components/common/Spacing/Spacing';
+import { CardCVCNumber, CardExpirationDate, CardNumber, CardPreview, Spacing } from '@/components';
 import { MASTER_CARD_CONDITIONS, VISA_CARD_CONDITIONS } from '@/constants';
 import { CardType, SequenceType, DateType } from '@/types';
 import { Global } from '@emotion/react';
