+++ src/components/CardExpirationDate/CardExpirationDate.tsx
@@ -1,8 +1,4 @@
-import Title from '@/components/common/Title/Title';
-import Label from '@/components/common/Label/Label';
-import Input from '@/components/common/Input/Input';
-import Spacing from '@/components/common/Spacing/Spacing';
-import ErrorMessage from '@/components/common/ErrorMessage/ErrorMessage';
+import { Title, Label, Input, Spacing, ErrorMessage } from '@/components/common';
 import { Dispatch, SetStateAction } from 'react';
 import { ERROR_MESSAGE } from '@/constants';
 import { DateType } from '@/types';
