+++ src/components/index.ts
@@ -1,0 +1,10 @@
+export { default as CardCVCNumber } from './CardCVCNumber/CardCVCNumber';
+export { default as CardExpirationDate } from './CardExpirationDate/CardExpirationDate';
+export { default as CardNumber } from './CardNumber/CardNumber';
+export { default as CardPreview } from './CardPreview/CardPreview';
+
+export { default as Spacing } from './common/Spacing/Spacing';
+export { default as Title } from './common/Title/Title';
+export { default as Label } from './common/Label/Label';
+export { default as Input } from './common/Input/Input';
+export { default as ErrorMessage } from './common/ErrorMessage/ErrorMessage';
